Create rich-vs-plain-text.md
# Rich Text vs Plain Text

Microsoft Word or Google Docs is Rich Text, whereas Notepad is Plain Text.

Rich Text is human friendly. Plain Text is computer friendly.

In Rich Text we can apply styles. (font type, font size, color, bold, italic,
letter spacing etc). It is pleasent for human eyes as the important part of the
text are prominently highlighted. But they are not computer friendly. We know
that in computer everything is represented as numbers (0s and 1s). When we type
'A', the character will be mapped with a binary number. But in Rich Text, the
computer has to read more information with respect to styles. On the other hand,
computers can easily read the 'Plain Text'.

We want to keep all our content as plain text. But by using simple markdown
language, we can get some of the common Rich Text features in Plain Text files.
Markdown files have the extension '.md'.

# MarkDown

## Title

```
# Top level title (usually Playlist name)
## Second level title (using Activity label)
### Third level title (used within an activity)
(We don't use more than three levels of titles)
```

## Code

To give a set of sentences, we enclose them between three back-ticks (```)

https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet